Python-Basics/EXERCISE_1/ex-1-5d/pyproject.toml
[tool.poetry]
name = "ex-1-5d"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

packages = [{include = "hello"}]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
